fb_messenger_bot/requirements.txt
undetected-chromedriver
python-dotenv
